% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_calling.R
\name{arg_calling}
\alias{arg_calling}
\title{Pull out an line, expression, or argument calling a specified function}
\usage{
arg_calling(ex, ..., n = 1L, message = "call to function")
}
\arguments{
\item{ex}{An call to be checked}

\item{...}{unquoted function names}

\item{n}{Look for the nth passing argument}

\item{message}{character string message on failure}
}
\description{
Find an argument
}
\examples{
ex <- for_checkr(quote(15 * sin(53 * pi / 180)))
ex2 <- for_checkr(quote(yy <- sin(3))) # at the top level
line_calling(ex, sin)
# look only in arguments: the top-level function doesn't count
arg_calling(ex, sin)
arg_calling(ex2, sin)
# But line_calling() will find it at the top level.
line_calling(ex2, sin)

}
