% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_choice}
\alias{assert_choice}
\title{Asserts that input is a single element of a set}
\usage{
assert_choice(input, set, null.ok = FALSE)
}
\arguments{
\item{input}{The input to check}

\item{set}{The full set}

\item{null.ok}{If set to TRUE, then null element is fine}
}
\description{
Asserts that input is a single element of a set
}
