% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_factor}
\alias{assert_factor}
\title{Asserts that input is factor}
\usage{
assert_factor(input, len = length(input), null_ok = FALSE)
}
\arguments{
\item{input}{Input that is being checked for being a factor}
}
\description{
Asserts that input is factor
}
