% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_file_type}
\alias{assert_file_type}
\title{Asserts that the input is of specified file type}
\usage{
assert_file_type(input, type)
}
\arguments{
\item{input}{Input that is checked for being specified filetype.}

\item{type}{File type to check for}
}
\description{
Asserts that the input is of specified file type
}
