% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_integer}
\alias{assert_integer}
\title{Asserts that input is integer}
\usage{
assert_integer(input, len = length(input), lower = -Inf, upper = Inf,
  null_ok = FALSE)
}
\arguments{
\item{input}{Input that is checked for being an integer}

\item{len}{The assumed length of the input}

\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{null_ok}{If true, will not assert null value}
}
\description{
Asserts that input is integer
}
