% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_logical}
\alias{assert_logical}
\title{Asserts that input is logical}
\usage{
assert_logical(input, len = length(input), null_ok = FALSE)
}
\arguments{
\item{input}{Input that is being checked for being a logical}
}
\description{
Asserts that input is logical
}
