% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Asserts that input is numeric}
\usage{
assert_numeric(input, len = length(input), lower = -Inf, upper = Inf,
  null_ok = FALSE)
}
\arguments{
\item{input}{Input that is checked for being numeric}

\item{len}{The assumed length of the input}

\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{null_ok}{If \code{TRUE}, will not assert null value}
}
\description{
Asserts that input is numeric
}
