% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_subset}
\alias{assert_subset}
\title{Asserts that input is subset of a set}
\usage{
assert_subset(input, set, null.ok = FALSE)
}
\arguments{
\item{input}{The input to check for being a subset}

\item{set}{The full set}

\item{null.ok}{If set to TRUE, then null subset is fine}
}
\description{
Asserts that input is subset of a set
}
