% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_tidy_table}
\alias{assert_tidy_table}
\title{Asserts that input is a \code{\link[tibble]{data_frame}}}
\usage{
assert_tidy_table(input, cols = NULL)
}
\arguments{
\item{input}{The input to check}

\item{cols}{Character vector of expected columns. Defaults to NULL, i.e., no preferenses.}
}
\description{
Asserts that input is a \code{\link[tibble]{data_frame}}
}
