% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{expand_chain}
\alias{expand_chain}
\alias{expand_all_chains}
\alias{chain_elements}
\alias{is_chain}
\title{Expand a chain into a series of lines, each with \code{.} as the input.}
\usage{
expand_chain(ex)

expand_all_chains(ex)

chain_elements(ex)

is_chain(ex)
}
\arguments{
\item{ex}{A \code{checkr_result} object with just one line of code.}
}
\value{
A \code{checkr_result} object with one line for each of the functions in the chain.
}
\description{
Expand a chain into a series of lines, each with \code{.} as the input.
}
\details{
A magrittr chain consists of a sequence of function calls. Each function takes as input
the output of the function before it. (The first element of the chain may be an object
or a function call.) The \code{expand_} functions transform chains into a sequence of lines. Each such line
(except the first) will be a function with at least one of the inputs being denoted \code{.}. The value
of \code{.} for each line will be the object that is an input to that line.
}
\examples{
code <- for_checkr(quote({x <- 3 \%>\% sin( ) \%>\% cos(); x \%>\% sqrt() \%>\% log()}))
lineA <- line_chaining(code)
expand_chain(lineA)
expand_all_chains(code)
}
