% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{check_atomic}
\alias{check_atomic}
\title{Check Atomic}
\usage{
check_atomic(x, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Atomic
}
\examples{
check_atomic(1)
check_atomic(character(0))
check_atomic(NULL)
check_atomic(Sys.Date())
check_atomic(Sys.time())
check_atomic(list(), error = FALSE)
x <- 1
names(x) <- "y"
check_atomic(x)
}
