% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classes.R
\name{check_class}
\alias{check_class}
\title{Check Object Class}
\usage{
check_class(
  x,
  valid_classes,
  specific = FALSE,
  cs_level = 1,
  msg = NULL,
  handler = "stop"
)
}
\arguments{
\item{x}{Object to be tested.}

\item{valid_classes}{\code{character}. Valid classes to check for.}

\item{specific}{\code{lgl}. Only match specific class?}

\item{cs_level}{\code{integer}. Indicater of how far up the
call-stack, the message should look for the call to report.
By default (1) check reports on parent function (
the function calling the check function). A level of 2 means
that check reports on the grandparent function. And so on.
See details.}

\item{msg}{\code{character} of length 1. A custom
message to overwrite the default assertthat message.}

\item{handler}{\code{character} of length 1. One of
'stop' or 'warning'. Should failed check result in
errors or warnings?}
}
\description{

}
