% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{check_count}
\alias{check_count}
\title{Check Count}
\usage{
check_count(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if object is a non-negative integer.
}
\examples{
check_count(-1L, error = FALSE)
check_count(1L, error = FALSE)
check_count(1, error = FALSE)
}
