% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
check_data(x, values, min_nrow = 0L, max_nrow = .Machine$integer.max,
  exclusive = FALSE, order = FALSE, key = character(0),
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{values}{An optional character vector specifying the column names or a named list specifying the column names and values.}

\item{min_nrow}{A count of the minimum number of rows}

\item{max_nrow}{A count of the maximum number of rows.}

\item{exclusive}{A flag indicating whether other columns are permitted.}

\item{order}{A flag indicating whether the columns have to occur in the same order as values.}

\item{key}{A character vector of the columns that represent a unique key.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Data
}
\examples{
z <- data.frame(
 Count = c(0L, 3L, 3L, 0L, NA), 
  Longitude = c(0, 0, 90, 90, 180), 
  Latitude = c(0, 90, 90.2, 100, -180),
  Type = factor(c("Good", "Bad", "Bad", "Bad", "Bad"), levels = c("Good", "Bad")),
  Extra = TRUE,
  Comments = c("In Greenwich", "Somewhere else", "I'm lost", 
    "I didn't see any", "Help"), 
  stringsAsFactors = FALSE)

check_data(z, values = list(
  Count = 1,
  Extra = NA,
  Latitude = c(45, 90)
  ), exclusive = TRUE, order = TRUE, min_nrow = 10L, key = "Longitude", error = FALSE)
}
