% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty.R
\name{check_empty}
\alias{check_empty}
\title{check_empty}
\usage{
check_empty(USER_CODE, message = "there is no content.")
}
\arguments{
\item{USER_CODE}{list object}

\item{message}{string}
}
\value{
list object
}
\description{
Checks whether the USER_CODE object after submission has any content at all.
Otherwise this will result in an uncaught error.
}
\details{
The function checks if any valid lines are found. If not a flag and the message are set.
}
\author{
Reinhard Simon
}
