% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{check_environment}
\alias{check_environment}
\title{Check Environment}
\usage{
check_environment(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or 
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if x is an environment.
}
\examples{
check_environment(1, error = FALSE)
check_environment(.GlobalEnv, error = FALSE)
}
