% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgl.R
\name{check_flag_na}
\alias{check_flag_na}
\title{Check Flag or NA}
\usage{
check_flag_na(x, coerce = TRUE, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{coerce}{A flag indicating whether to drop attributes including names.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if x is a flag or NA (missing logical scalar).
}
\details{
Useful when using flag to pass one of three options.
}
\examples{
check_flag_na(1, error = FALSE)
check_flag_na(FALSE)
check_flag_na(NA)
check_flag_na(NA_character_, error = FALSE)
check_flag_na(c(NA, NA), error = FALSE)
}
