% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_learnr.R
\name{check_for_learnr}
\alias{check_for_learnr}
\title{Interface for checking learnr exercises.}
\usage{
check_for_learnr(label = NULL, user_code = NULL, solution_code = NULL,
  check_code = NULL, envir_result = NULL, evaluate_result = NULL, ...,
  debug = FALSE)
}
\arguments{
\item{label}{argument passed by learnr system}

\item{user_code}{ditto}

\item{solution_code}{ditto}

\item{check_code}{ditto}

\item{envir_result}{ditto}

\item{evaluate_result}{ditto}

\item{...}{ditto}

\item{debug}{development flag to turn on logging of the information sent by learnr. This
must be set at compile time, it's not for users.}
}
\description{
This is the interface from learnr to checkr. You don't call this function directly: that will be
done by learnr. Instead, you have to tell learnr to use this function. Do this with a directive
in the setup chunk of the learnr document: \code{tutorial_options(exercise.checker = checkr2::check_for_learnr)}
}
\examples{
# as it would be called from the learnr system ...
check_for_learnr(envir_result = 3, label = "first", user_code = "sin(pi)",
  check_code = 'ex <- for_checkr(USER_CODE)
                t1 <- line_where(ex, F == sin, message="Please use the sin function.")
                a1 <- arg_number(t1, 1); check(a1, passif(V == pi, "Right-oh!"))'
)

#'
}
