% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_function}
\alias{check_function}
\title{Check Function}
\usage{
check_function(x, nargs = c(0L, .Machine$integer.max),
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{nargs}{A count of the number of arguments or count range of the minimum and maximum number of arguments.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if x is a function.
}
\examples{
check_function(character, error = FALSE)
check_function(character, nargs = 0L, error = FALSE)
}
