% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherits.R
\name{check_inherits}
\alias{check_inherits}
\title{Check Inherits}
\usage{
check_inherits(x, class, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A string of the class x should inherit from.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if an object inherits from a class.
}
\examples{
check_inherits(list(), "list")
check_inherits(list(), "numeric", error = FALSE)
}
\seealso{
\code{\link[=check_classes]{check_classes()}}
}
