% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{check_length1}
\alias{check_length1}
\title{Check Length One}
\usage{
check_length1(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or 
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether x is an object of length 1.
}
\examples{
check_length1(2)
check_length1(1:2, error = FALSE)
check_length1(NULL, error = FALSE)
check_length1(list(), error = FALSE)
}
