% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{check_linearity}
\alias{check_linearity}
\title{check_linearity}
\usage{
check_linearity(model, independent, dependent)
}
\arguments{
\item{model}{A linear regression model object ('lm')}

\item{independent}{The independent variable to test}

\item{dependent}{The dependent variable to test}
}
\value{
A scatter plot
}
\description{
check_linearity
}
