% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{check_list_class}
\alias{check_list_class}
\title{Check Vector Class}
\usage{
check_list_class(x, values, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector specifying the class.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Vector Class
}
\examples{
check_vector_class(2, 1, error = FALSE)
check_vector_class(2L, 1, error = FALSE)
check_vector_class(list(x = 1), list(), error = FALSE)
}
