% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{check_list_values}
\alias{check_list_values}
\title{Check List Values}
\usage{
check_list_values(x, values, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{values}{A named list specifying the values.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check List Values
}
\examples{
check_list_values(list(x = 2), values = list(x = 1), error = FALSE)
}
