% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{check_logit}
\alias{check_logit}
\title{check_logit}
\usage{
check_logit(model)
}
\arguments{
\item{model}{a logit model object to test}
}
\value{
A list containing:
\itemize{
  \item the difference between the null deviance and model deviance
  \item the difference in degrees of freedom
  \item the chi-squared probability of the deviance statistic
  \item Cox and Snell pseudo R-squared
  \item Nagelkerke pseudo R-squared; and
  \item Hosmer pseudo R-squared.
}
}
\description{
Check logistic regression statistics and assumptions
}

