% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlevels.R
\name{check_nlevels}
\alias{check_nlevels}
\title{Check nlevels}
\usage{
check_nlevels(x, min_nlevels = 1L, max_nlevels = 2147483647L,
  x_name = substitute(x), error = getOption("checkr.error", TRUE))
}
\arguments{
\item{x}{The data to check.}

\item{min_nlevels}{A count of the minimum number of rows.}

\item{max_nlevels}{A count of the maximum number of rows.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check nlevels
}
\examples{
check_nlevels(factor(1), error = FALSE)
check_nlevels(factor(1), min_nlevel = 2, error = FALSE)
}
