% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{check_nrow}
\alias{check_nrow}
\title{Check nrow}
\usage{
check_nrow(x, min_nrow = 1L, max_nrow = 2147483647L,
  x_name = substitute(x), error = getOption("checkr.error", TRUE))
}
\arguments{
\item{x}{The data to check.}

\item{min_nrow}{A count of the minimum number of rows.}

\item{max_nrow}{A count of the maximum number of rows.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check nrow
}
\examples{
check_nrow(data.frame(x = 1), error = FALSE)
check_nrow(data.frame(x = integer(0)), error = FALSE)
check_nrow(data.frame(x = 1:2), max_nrow = 1, error = FALSE)
}
