% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{check_null}
\alias{check_null}
\title{Check NULL}
\usage{
check_null(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check NULL
}
\examples{
check_null(1, error = FALSE)
check_null(NULL, error = FALSE)
}
