% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{check_props}
\alias{check_props}
\title{Check Proportions}
\usage{
check_props(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if x is proportions vector - non-missing dbls between 0 and 1 that sum to 1.
}
