% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{check_scalar}
\alias{check_scalar}
\title{Check Scalar}
\usage{
check_scalar(x, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Scalar
}
\examples{
check_scalar(2)
check_scalar(character(0), error = FALSE)
check_scalar(NULL, error = FALSE)
}
