% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnamed.R
\name{check_unnamed}
\alias{check_unnamed}
\title{Check Unnamed}
\usage{
check_unnamed(x, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Unnamed
}
\examples{
check_unnamed(2, error = FALSE)
x <- 1
names(x) <- "y"
check_unnamed(x, error = FALSE)
}
