% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{check_vector_nas}
\alias{check_vector_nas}
\title{Check NAs}
\usage{
check_vector_nas(x, values, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector with no, some or all NAs.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check NAs
}
\examples{
check_nas(NA, 1, error = FALSE)
check_nas(NA, NA, error = FALSE)
check_nas(1, NA, error = FALSE)
}
