% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{check_vector_values}
\alias{check_vector_values}
\title{Check Vector Values}
\usage{
check_vector_values(x, values, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector specifying the values.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Vector Values
}
\examples{
check_vector_values(2, c(3,4), error = FALSE)
check_vector_values(2, c(1,3), error = FALSE)
check_vector_values(2, c(1,3,4), error = FALSE)
check_vector_values(c(0L, 1L, 11L), 1:10, error = FALSE)
check_vector_values("t1", c("t0", "t2"), error = FALSE)
check_vector_values("t3", c("t0", "t2"), error = FALSE)
}
