% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkor.R
\name{checkor}
\alias{checkor}
\title{Check OR}
\usage{
checkor(..., error = TRUE)
}
\arguments{
\item{...}{The checks to check.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if all checks fails.}
}
\value{
An invisible flag indicating whether at least one check passes (if it doesn't throw an error).
}
\description{
Checks that at least one check passes.
}
\examples{
checkor(check_null(NULL), check_null(1), error = FALSE)
checkor(check_null(1), check_null(1), error = FALSE)
checkor(check_null(1), check_null(2), error = FALSE)
}
