% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_deparse}
\alias{chk_deparse}
\title{Deparse}
\usage{
chk_deparse(x)
}
\arguments{
\item{x}{A substituted object to deparse}
}
\value{
A string
}
\description{
\code{chk_deparse} is a wrapper on \code{\link{deparse}} that
sets a missing value to be "NA"
}
\examples{
chk_deparse(1^2)
}
\seealso{
\code{\link{deparse}}
}
