% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_fail}
\alias{chk_fail}
\title{Fail}
\usage{
chk_fail(..., error)
}
\arguments{
\item{...}{The message.}

\item{error}{A flag indicating whether to return an error (the default) or a warning.}
}
\description{
Fail
}
