% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_max_integer}
\alias{chk_max_integer}
\title{Max Integer}
\usage{
chk_max_integer()
}
\value{
An integer of the maximum integer value for the system.
}
\description{
Max Integer
}
\examples{
chk_max_integer()
}
