% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_min_integer}
\alias{chk_min_integer}
\title{Min Integer}
\usage{
chk_min_integer()
}
\value{
An integer of the minimum integer value for the system.
}
\description{
Min Integer
}
\examples{
chk_min_integer()
}
