% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_tiny_dbl}
\alias{chk_tiny_dbl}
\title{Tiny Positive Double}
\usage{
chk_tiny_dbl()
}
\value{
An dbl of the tiniest positive numeric value for the system.
}
\description{
Tiny Positive Double
}
\examples{
chk_tiny_dbl()
}
