% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{\%same_as\%}
\alias{\%same_as\%}
\alias{\%not_same_as\%}
\alias{same_val}
\alias{same_name}
\alias{`same_val`}
\alias{checks}
\alias{whether}
\alias{its}
\alias{two}
\alias{arguments}
\alias{have}
\alias{the}
\alias{same}
\alias{value.}
\alias{One}
\alias{or}
\alias{both}
\alias{of}
\alias{can}
\alias{be}
\alias{a}
\alias{name,}
\alias{in}
\alias{which}
\alias{case}
\alias{name}
\alias{is}
\alias{evaluated}
\alias{to}
\alias{produce}
\alias{`same_name`}
\alias{correspond}
\alias{name.}
\alias{Either}
\alias{character}
\alias{string}
\alias{corresponding}
\alias{`\%same_as\%`}
\alias{\%not_same_as\%}
\alias{same_name}
\title{Comparison operators for bindings}
\usage{
e1 \%same_as\% e2

e1 \%not_same_as\% e2

same_name(e1, e2)
}
\arguments{
\item{e1}{an expression,}

\item{e2}{another expression}
}
\description{
Comparison operators for bindings
}
