% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{contains_regexp}
\alias{contains_regexp}
\title{Checks if input contains regexp \code{pattern}}
\usage{
contains_regexp(input, pattern)
}
\arguments{
\item{input}{The input to check}

\item{pattern}{The pattern}
}
\description{
Checks if input contains regexp \code{pattern}
}
