% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_on_false.R
\name{error_on_false}
\alias{error_on_false}
\title{Invoke Formatted Error on FALSE check}
\usage{
error_on_false(
  check,
  x,
  var_name,
  input_name,
  expected,
  allow_null,
  allow_na,
  call,
  class
)
}
\arguments{
\item{check}{A length 1 boolean as a result of a predicate expression.}

\item{x}{the object to check}

\item{var_name}{the name the user gave the input}

\item{input_name}{the name of the input to check}

\item{expected}{a string describing what the input should be}

\item{allow_null}{allow check to pass if x is NULL}

\item{allow_na}{allow check to pass if any element of x is NA}

\item{call}{the call environment}

\item{class}{the class of the error}
}
\value{
Called for side-effect. Returns TRUE/FALSE invisibly.
}
\description{
Given a length 1 boolean \code{check}, invokes a
formatted error if \code{check} is FALSE.
}
