% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkr.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a Checkr object}
\usage{
evaluate(chckr, x, level = -1)
}
\arguments{
\item{chckr}{Checkr object}

\item{x}{object to check}

\item{level}{deparse level to determine original name of x}
}
\value{
NA_character or character
}
\description{
Evaluate executes the provided isvalid() function of a given Checkr object.
If it returns TRUE, NA is returned, otherwiese a string generated from
message().
}
