% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_tests.R
\name{formula_equals}
\alias{formula_equals}
\alias{formula_equals,}
\alias{two_sided,}
\alias{lhs_equals,}
\alias{rhs_equals,}
\alias{rhs_contains}
\alias{two_sided}
\alias{lhs_equals}
\alias{rhs_equals}
\alias{rhs_contains}
\title{Testing functions for formulas}
\usage{
formula_equals(F, target)

two_sided(F)

lhs_equals(F, ...)

rhs_equals(F, ...)

rhs_contains(F, ...)
}
\arguments{
\item{F}{a formula from a student submission.}

\item{target}{the desired attributes of the formula}

\item{...}{expressions for the permissible parts of the formula}
}
\description{
Provide facilities for checking the layout of formulas. Use these
with an input that is the \emph{value} of a formula, that is, with a V binding rather
than an EX binding.
}
\examples{
CODE <- for_checkr(quote({
   data(mtcars, package = "datasets")
   mod <- lm(mpg ~ hp * wt + cyl, data = mtcars)
   }))
fa <- formula_arg(line_calling(CODE, lm),
  failif(! two_sided(V), "Modeling formulas should have two sides."),
  failif(! lhs_equals(V, mpg), "The response variable should be mileage."))
# or, another style for the same thing ...
check(formula_arg(line_calling(CODE, lm)),
  failif(! two_sided(V), "Modeling formulas should have two sides."),
  failif(! lhs_equals(V, mpg), "The response variable should be mileage."))
}
