% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcheck.R
\name{function_name}
\alias{function_name}
\title{Get the name from a passed function, which may be a validated function or just a block.}
\usage{
function_name(orig_function_name)
}
\arguments{
\item{orig_function_name}{call. A substituted call of the function.}
}
\description{
Get the name from a passed function, which may be a validated function or just a block.
}

