% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quickcheck.R
\name{function_test_objects}
\alias{function_test_objects}
\title{Create the necessary testing objects to quickcheck a function.}
\usage{
function_test_objects(...)
}
\arguments{
\item{fn}{function. A function to generate test objects for.}
}
\description{
Create the necessary testing objects to quickcheck a function.
}

