% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{get_prevalidated_fn}
\alias{get_prevalidated_fn}
\title{Get the pre-validated function that is wrapped in validations.}
\usage{
get_prevalidated_fn(...)
}
\arguments{
\item{fn}{validated_function. The function to get the pre-validated function for.}
}
\value{
a call containing the postconditions.
}
\description{
Get the pre-validated function that is wrapped in validations.
}

