% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{\%does_not_contain\%}
\alias{\%does_not_contain\%}
\title{Test if a list does not contain some elements of the desired class.}
\usage{
... \%does_not_contain\% NA
}
\arguments{
\item{match_list.}{The list to test for class of the elements.}

\item{expected_class.}{The name of the expected class to test.}
}
\value{
Boolean whether or not the match_list has no elements of the expected_class.
}
\description{
Test if a list does not contain some elements of the desired class.
}
\examples{
  list(1, 2, 3) \%does_not_contain\% character
}

