% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{\%is\%}
\alias{\%is\%}
\alias{\%is_not\%}
\title{Test for class membership}
\usage{
match_object \%is\% expected_class

match_object \%is_not\% expected_class
}
\arguments{
\item{match_object}{ANY. The object to test for class.}

\item{expected_class}{character. The name of the expected class.}
}
\value{
Boolean whether or not the match_object is the expected_class.
}
\description{
Test for class membership

Test whether a match object is not a member of a particular class.
}
\examples{
  1 \%is\% numeric
  1.0 \%is\% double
  1L \%is\% integer
  iris \%is\% dataframe
  c("a", "b", "c") \%is\% vector
  "pizza" \%is\% simple_string
  list(a = "pizza", b = "pie") \%is\% c("character", "list")
}
