% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{\%within\%}
\alias{\%within\%}
\title{Define if a number is within a certain range.}
\usage{
num \%within\% range
}
\arguments{
\item{num}{numeric. The number to check.}

\item{range}{numeric. A vector with one number specifying the lower-bound and another number specifying the upper-bound.}
}
\description{
Define if a number is within a certain range.
}

