% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_empty_submission.R
\name{if_empty_submission}
\alias{if_empty_submission}
\title{Test for an empty submission}
\usage{
if_empty_submission(ex, message = "Give me something to start with!")
}
\arguments{
\item{ex}{A checkr_result object, typically straight from \code{for_checkr}}

\item{message}{The message to give if \code{ex} has no code associated with it.}
}
\description{
Handles a special case when a submission is empty. This function
provides a way to through a graceful error, perhaps with a message telling
the student how to start.
}
\examples{
code <- for_checkr("")
t1 <- if_empty_submission(code, message = "Start out with something. Say 2+2.")
}
