% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate.R
\name{installed_dataframes}
\alias{installed_dataframes}
\title{Get all the user-installed dataframes through data()}
\usage{
installed_dataframes()
}
\description{
Get all the user-installed dataframes through data()
}

