% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_string.R
\name{is.simple_string}
\alias{is.simple_string}
\title{Tests whether a string is simple.}
\usage{
is.simple_string(string)
}
\arguments{
\item{string}{character.}
}
\value{
a boolean whether or not string is simple string.
}
\description{
A simple string is an R object that is a length-1 vector of non-empty characters.
}
\examples{
  is.simple_string("pizza")              # true
  is.simple_string(c("pizza", "apple"))  # false
  is.simple_string(iris)                 # false
  is.simple_string(NA)                   # false
}
