% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_exports_checked.R
\name{is.validated_function}
\alias{is.validated_function}
\title{Determine whether a function is checked with checkr.}
\usage{
is.validated_function(fn)
}
\arguments{
\item{fn}{function. The function to check.}
}
\description{
Determine whether a function is checked with checkr.
}
