% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_character}
\alias{is_character}
\title{Checks if input is character}
\usage{
is_character(input, len = length(input))
}
\arguments{
\item{input}{The input to check}

\item{len}{The assumed length of input}
}
\description{
Checks if input is character
}
