% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_factor}
\alias{is_factor}
\title{Checks if input is factor}
\usage{
is_factor(input, len = length(input))
}
\arguments{
\item{input}{The input to check}

\item{len}{The assumed length of input}
}
\description{
Checks if input is factor
}
