% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_file_type}
\alias{is_file_type}
\title{Checks if input is of filetype \code{type}}
\usage{
is_file_type(input, type)
}
\arguments{
\item{input}{The input to check}

\item{type}{The filetype}
}
\description{
Checks if input is of filetype \code{type}
}
