% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_integer}
\alias{is_integer}
\title{Checks if input is integer}
\usage{
is_integer(input, len = length(input), lower = -Inf, upper = Inf)
}
\arguments{
\item{input}{The input to check}

\item{len}{The assumed length of input}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Checks if input is integer
}
