% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_logical}
\alias{is_logical}
\title{Checks if input is logical}
\usage{
is_logical(input, len = length(input))
}
\arguments{
\item{input}{The input to check}

\item{len}{The assumed length of input}
}
\description{
Checks if input is logical
}
