% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_numeric}
\alias{is_numeric}
\title{Checks if input is numeric}
\usage{
is_numeric(input, len = length(input), lower = -Inf, upper = Inf)
}
\arguments{
\item{input}{The input to check}

\item{len}{The assumed length of input}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Checks if input is numeric
}
