% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_subset}
\alias{is_subset}
\title{Checks if input is  a subset of the set}
\usage{
is_subset(input, set)
}
\arguments{
\item{input}{The input to check}

\item{set}{The full set}
}
\description{
Checks if input is  a subset of the set
}
