% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_tidy_table}
\alias{is_tidy_table}
\title{Checks if input is a \code{\link[tibble]{data_frame}}}
\usage{
is_tidy_table(input, cols = NULL)
}
\arguments{
\item{input}{The input to check}

\item{cols}{Character vector of expected columns. Defaults to NULL, i.e., no preferenses.}
}
\description{
Checks if input is a \code{\link[tibble]{data_frame}}
}
