% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_tidy_topic_state}
\alias{is_tidy_topic_state}
\title{Checks if input is a \code{\link[tibble]{tibble}}
having columns 'doc', 'type', 'topic'}
\usage{
is_tidy_topic_state(state)
}
\arguments{
\item{The}{input to check}
}
\description{
Checks if input is a \code{\link[tibble]{tibble}}
having columns 'doc', 'type', 'topic'
}
