% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{list_classes}
\alias{list_classes}
\title{Get all the classes within a list.}
\usage{
list_classes(object)
}
\arguments{
\item{object}{ANY. The object to check classes for.}
}
\description{
Get all the classes within a list.
}

