% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misconception.R
\name{misconception}
\alias{misconception}
\title{Test for a misconceived statement}
\usage{
misconception(ex, pattern, message)
}
\arguments{
\item{ex}{A checkr result to use as input}

\item{pattern}{A code pattern as found by any of the functions returning a checkr_result object.}

\item{message}{The message to give if the \code{pattern} is found.}
}
\description{
Sometimes you expect a specific code pattern which might reflect a common misconception.
\code{misconception()} let's you mark a pattern as such, causing the test to fail if the pattern
is found.
}
\examples{
code <- for_checkr(quote({theta <- 53; x <- sin(theta)}))
misconception(code, line_where(code,V == 53), message = "Angle should be in degrees, not radians.")
misconception(code, line_calling(code, sin), message = "For the x-coordinate, use cos().")

}
