% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moustache.R
\name{moustache}
\alias{moustache}
\title{evaluate expressions in a string using a particular environment}
\usage{
moustache(string, bindings = rlang::env_parent())
}
\arguments{
\item{string}{a character string which presumably contains some moustaches
referring to objects found in the bindings environment.}

\item{bindings}{an environment or list in which the objects moustached in \code{string} are defined.}
}
\description{
evaluate expressions in a string using a particular environment
}
\examples{
checkr2:::moustache("hello")
checkr2:::moustache("Three plus five is {{3+5}}.")
}
