% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete_funs.R
\name{capture.code}
\alias{capture.code}
\alias{find_assignment}
\alias{find_call}
\alias{find_constants}
\alias{find_formula}
\alias{find_names}
\alias{all_tests}
\alias{any_test}
\alias{arg_is}
\alias{checkr_tutor}
\alias{check_argument}
\alias{check_function_calls}
\title{Defunct functions from checkr (v 0.1)}
\usage{
capture.code(...)

find_assignment(...)

find_call(...)

find_constants(...)

find_formula(...)

find_names(...)

all_tests(...)

any_test(...)

arg_is(...)

checkr_tutor(...)

check_argument(...)

check_function_calls(...)
}
\arguments{
\item{...}{whatever arguments the original functions took}
}
\description{
These functions were in version 0.1 of checkr, but as of
version 0.3 they are not part of the system. If you are using them,
you are using an old version of checkr. If you want to continue using them,
install that old version: \code{devtools::install_github("dtkaplan/checkr-v-0.1")}
}
