% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_exports_checked.R
\name{package_exports_checked}
\alias{package_exports_checked}
\title{Checks if all exported functions in the package are checked using checkr.}
\usage{
package_exports_checked(package)
}
\arguments{
\item{package}{character. The name of the package to check.}
}
\description{
Checks if all exported functions in the package are checked using checkr.
}

