% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{params}
\alias{params}
\title{Parameters for checkr functions}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or
immediately generate an informative message if the check fails.}

\item{pattern}{A string of the regular expression.}

\item{y_name}{A string of the name of the object y.}

\item{values}{NULL or a vector specifying the values.}

\item{exclusive}{A flag indicating whether other columns are not permitted.}

\item{by}{A character vector or named character vector of the columns to join by.}

\item{colnames}{A character vector of the column names.}

\item{na_distinct}{A flag specifying whether missing values should be considered distinct.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{only}{A flag indicating whether only the actual values are permitted.
It only affects values with two or less non-missing elements.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{length}{A flag indicating whether x should have elements (versus no elements)
or a missing value indicating no requirements or a count or count range of the
number of elements or a count vector of the permitted number of elements.}

\item{complete}{A flag indicating whether all the possible names have to be represented.}

\item{coerce}{A flag indicating whether to coerce a scalar integer to a dbl and drop attributes including names.}

\item{tzone}{A string of the time zone.}

\item{strict}{A flag indicating whether all the objects must have identical classes or just share one or more classes.}

\item{recursive}{A flag indicating whether the check should be applied recursively.}

\item{ncol}{A flag indicating whether x should have columns (versus no columns) or a missing value indicating no requirements or a count or count range of the number of columns.}

\item{nlevels}{A flag indicating whether x should have elements (versus no elements) or a missing value indicating no requirements or a count or count range of the number of elements.}

\item{nrow}{A flag indicating whether x should have rows (versus no rows) or a missing value indicating no requirements or a count or count range of the number of rows.}

\item{nargs}{A count of the number of arguments or count range of the minimum and maximum number of arguments.}

\item{levels}{A character vector of the levels.}
}
\description{
Descriptions of the parameters for the checkr functions.
}
\keyword{internal}
