% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_check.R
\name{parse_check}
\alias{parse_check}
\title{Look for errors at parse time}
\usage{
parse_check(user_code)
}
\arguments{
\item{user_code}{a character string containing the code to parse. Note, the input
cannot be a checkr_result, since constructing one of those requires the code to be
executed.}
}
\description{
Parses the code (without evaluating it) to look for parse errors. If any are found,
a somewhat friendly message is returned.
}
\examples{
code <- "lm(mpg ~ hp, \\ndata == mtcars); \\nplot(1:10)\\n x <- f[b]\\n y <- x^2\\n\\n z = cos(yy * xx^2)"
cat(parse_check(code)$message)
code2 <- "x <- 3\\ny <- x^2\\n__blank__ <- 2"
cat(parse_check(code2)$message)
code3 <- "x["
cat(parse_check(code3)$message)
}
