% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{postconditions}
\alias{postconditions}
\title{Get the stated postconditions of a validated function.}
\usage{
postconditions(fn)
}
\arguments{
\item{fn}{validated_function. The function to get the postconditions for.}
}
\value{
a call containing the postconditions.
}
\description{
Get the stated postconditions of a validated function.
}
