% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensure.R
\name{preconditions}
\alias{preconditions}
\title{Get the stated preconditions of a validated function.}
\usage{
preconditions(fn)
}
\arguments{
\item{fn}{validated_function. The function to get the preconditions for.}
}
\value{
a call containing the preconditions.
}
\description{
Get the stated preconditions of a validated function.
}

