% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present}
\alias{present}
\title{Tests whether an argument to a function is present.}
\usage{
present(...)
}
\arguments{
\item{...}{list. The list of things to check for presence.}
}
\description{
This function is the opposite of missing.
}

