% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.examiner_result.R
\name{print.examiner_result}
\alias{print.examiner_result}
\title{Print a examiner_result}
\usage{
\method{print}{examiner_result}(x, ...)
}
\arguments{
\item{x}{a examiner_result object to be printed}

\item{...}{additional args (not used here)}
}
\description{
Print a examiner_result
}
