% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{print.validated_function}
\alias{print.validated_function}
\title{Print validated functions more clearly.}
\usage{
\method{print}{validated_function}(x, ...)
}
\arguments{
\item{x}{function. The function to print.}

\item{...}{list. Additional arguments to pass to print.}
}
\description{
Print validated functions more clearly.
}
