% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcheck.R
\name{print_args}
\alias{print_args}
\title{Print function arguments}
\usage{
print_args(x)
}
\arguments{
\item{x}{ANY. The object to print args for.}
}
\description{
Print function arguments
}
\examples{
l <- list(x = seq(3), y = seq(4))
print_args(l)
}

