% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_function_contents.R
\name{print_function_contents}
\alias{print_function_contents}
\title{Print a function with numbered lines.}
\usage{
print_function_contents(fun, just_the_body = TRUE, from_file = NULL)
}
\arguments{
\item{fun}{Name of the function, e.g. \code{lm}.}

\item{just_the_body}{If \code{TRUE}, show just the body of the function. Otherwise, include
the function definition.}

\item{from_file}{name of a .R file. If the function resides in a .R file, giving \code{from_file}
will cause that .R file to be sourced.}
}
\description{
Many checkr statements will be packaged into functions with just one input: \code{USER_CODE}.
In narrative documentation for checkr, it can be nice to print these with lines numbered, so
that the narrative can refer to them. There's no reason to use this function in a learnr tutorial,
unless that tutorial is about checkr itself.
}
\details{
Won't work for primitives, but then you'll never be writing primitives.
}
