% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcheck.R
\name{quickcheck}
\alias{quickcheck}
\title{Quickcheck a function.}
\usage{
quickcheck(fn, postconditions = NULL, verbose = TRUE, testthat = TRUE)
}
\arguments{
\item{fn}{function. A function to randomly check postconditions for.}

\item{postconditions}{list. Optional postconditions to quickcheck for.}

\item{verbose}{logical. Whether or not to announce the success.}

\item{testthat}{logical. Whether or not to run testthat.}
}
\value{
either TRUE if the function passed the quickcheck or FALSE if it didn't.
}
\description{
Tests a function with many automatically generated inputs, checking that stated
postconditions hold.
}
\details{
If given a function of class \code{validated_function}, the pre- and post-conditions can
be automatically inferred by the definition of the function. The test objects used to
test the function will be screened ahead of time to ensure they meet the preconditions.
}

