% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate.R
\name{random_matrix}
\alias{random_matrix}
\title{Generate a random matrix.}
\usage{
random_matrix()
}
\description{
A random matrix needs three random things...
A random width, a random height, and a random data
data should be a random assortment of integers, doubles, logicals, or characters, with
  all of them being the same class.
Because there are so many possible matricies, it seems easier to generate them on
demand rather than preallocate all possible matricies into OBJECTS.
We will then populate some random matricies onto OBJECTS for later use.
}

