% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{random_simple_strings}
\alias{random_simple_strings}
\title{Generate a random simple string (i.e., a length-1 non-empty vector of characters).}
\usage{
random_simple_strings(amount, chars = TRUE, utf8 = FALSE)
}
\description{
Generate a random simple string (i.e., a length-1 non-empty vector of characters).
}

