% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run test statements in the context of the bindings}
\usage{
run_tests(test_list, bindings, ex)
}
\arguments{
\item{test_list}{a list containing a set of tests}

\item{bindings}{the bindings resulting from a pattern match to a submission}

\item{ex}{the expression being tested (used for failure message)}
}
\description{
Run test statements in the context of the bindings
}
