% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_exports_checked.R
\name{should_be_checked}
\alias{should_be_checked}
\title{Determine whether a function ought to be checked with checkr.}
\usage{
should_be_checked(fn)
}
\arguments{
\item{fn}{function. The function to check.}
}
\description{
Functions should be checked as long as they have formals.
}

