% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ex.R
\name{n_args}
\alias{n_args}
\alias{func_of}
\title{Tests on expressions}
\usage{
n_args(ex)

func_of(ex)
}
\arguments{
\item{ex}{A straight expression, for instance as bound to \code{EX} in \code{line_where().}}
}
\description{
These functions can be applied to EX bindings as produced by \code{line_where()}.
They calculate something about the EX expression.
}
