% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate.R
\name{test_objects_}
\alias{test_objects_}
\title{Generates random R objects to be put into functions for testing purposes.}
\usage{
test_objects_()
}
\description{
Generates random R objects to be put into functions for testing purposes.
}

