% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{test_residuals}
\alias{test_residuals}
\title{test_residuals}
\usage{
test_residuals(df, residuals, which = "sd5")
}
\arguments{
\item{df}{the data frame used in the logistic regression model containing
the residuals}

\item{residuals}{a character name of the column containing the residuals}

\item{which}{which level to test, either 5\% (1.96 standard deviations) or
1% (2.58 standard deviations)}
}
\value{
the proportion of outlying cases. Ideally this should be less than 5%
for 1.96 standard deviations, or less than 1% for 2.58 standard devations.
}
\description{
Calculate what percentage of studentised or standardised residuals
are outliers, either +- 1.96 standard deviations or +- 2.58 standard
deviations
}
