% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{type_check}
\alias{type_check}
\alias{check_is_integerish}
\alias{check_is_numeric}
\alias{check_is_character}
\title{Type checks}
\usage{
check_is_integerish(
  x,
  allow_null = FALSE,
  allow_na = TRUE,
  .x_name = rlang::caller_arg(x),
  .call = rlang::caller_env(),
  .class = .error_class
)

check_is_numeric(
  x,
  allow_null = FALSE,
  allow_na = TRUE,
  .x_name = rlang::caller_arg(x),
  .call = rlang::caller_env(),
  .class = .error_class
)

check_is_character(
  x,
  allow_null = FALSE,
  allow_na = TRUE,
  .x_name = rlang::caller_arg(x),
  .call = rlang::caller_env(),
  .class = .error_class
)
}
\arguments{
\item{x}{the object to check}

\item{allow_null}{allow check to pass if x is NULL}

\item{allow_na}{allow check to pass if x contains na}

\item{.x_name}{the name of x in the parent environment}

\item{.call}{The caller environment, used to get the name of the function that called the check}

\item{.class}{The class of the error, defaults to 'checkr_error'}
}
\value{
Called for side-effect. Returns TRUE invisibly if the check passes.
Invisibly FALSE otherwise.
}
\description{
Performs type checks on an object, invoking a
formatted error message containing the calling environment,
if the object is not of the expected type.
}
\section{Functions}{
\itemize{
\item \code{check_is_integerish()}: Checks whether an object is an  \code{integerish}.
See the rlang-function \link[rlang]{is_integerish} for the implementation of
\code{integerish}.

\item \code{check_is_numeric()}: Checks whether an object is a numeric vector.

\item \code{check_is_character()}: Checks whether an object is a character vector. Uses
the character definition from rlang. See \link[rlang]{is_character}.

}}
\examples{
library(checkR)
check_is_integerish(1)
check_is_numeric(2L)
check_is_numeric(2.5)

check_is_character("3")
}
