% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensure.R, R/validate.R
\name{validate}
\alias{validate}
\title{Require certain things are true about your function call.}
\usage{
validate(...)

validate(...)
}
\arguments{
\item{...}{list. A list of conditions to check.}

\item{param(s)}{type. Description of param}
}
\value{
Return Object Description

Either TRUE or stops with a list of errors.
}
\description{
Require certain things are true about your function call.

Validate checks that certain facts about the inputs of your parameters are true.
}
\examples{
\dontrun{
  a <- rnorm(100)
  b <- a*.7 + rnorm(100)
  plot(a, b)
  # If wrapped in /code{dontrun} than also sample of return object if possible
}
validate(1 == 1, "a" \%is\% character, length(c(1, 2, 3)) == 3)
}

