% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_character}
\alias{assert_character}
\title{Assets that the input is character.}
\usage{
assert_character(input, len = length(input), null_ok = FALSE)
}
\arguments{
\item{input}{Input that is checked for being a character}

\item{len}{The assumed length of the input}

\item{null_ok}{If \code{TRUE}, will not assert null value}
}
\description{
Assets that the input is character.
}
