% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_check.R
\name{check}
\alias{check}
\title{check helper}
\usage{
check(x, var_name, input_name, func, expected, call, class)
}
\arguments{
\item{x}{the variable to check}

\item{var_name}{the variable name the user gave the input}

\item{input_name}{the name of the input we are checking}

\item{func}{the function performing the check}

\item{expected}{a string describing what the input should be}

\item{call}{the call environment}

\item{class}{the class of the error}
}
\value{
invisible true/falses
}
\description{
check helper
}
