% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check That a Set of Conditions are True}
\usage{
check(..., cs_level = 1, msg = NULL, handler = "stop")
}
\arguments{
\item{...}{logical expressions to evaluate.
See \link[assertthat]{assert_that}.}

\item{cs_level}{\code{integer}. Indicater of how far up the
call-stack, the message should look for the call to report.
By default (1) check reports on parent function (
the function calling the check function). A level of 2 means
that check reports on the grandparent function. And so on.
See details.}

\item{msg}{\code{character} of length 1. A custom
message to overwrite the default assertthat message.}

\item{handler}{\code{character} of length 1. One of
'stop' or 'warning'. Should failed check result in
errors or warnings?}
}
\value{
TRUE if the check passed. Side-effect error or
warning otherwise.
}
\description{
Checks that a set of provided conditions are true and returns
informative error messages including a notice of where in the
call stack, the check failed. This is essentially a
wrapper around the assertthat package functionality, but allowing
the handler message to be prefixed with a function name in the
call stack. This means error or warning messages combine the
informative messaging from the assertthat package, while also
giving hints to the function call that contains the failed
check.
}
\details{
Under the hood \code{check()} is a wrapper around the
\code{see_if()} function from the assertthat package. This means
\code{check()} works similar to \code{assert_that()}, and
also with the utility functions of the assertthat package.

\strong{Call-stack function messaging}:

The addition that \code{check()} makes is that
messages start with a function call from
the call stack. This helps the user to identify in
which function a certain check failed. The default call
is the function call containing the check call
(cs_level = 1). That is, if a function A
contains a check that fails, calling \code{A()}, will
cause an error or warning with a message specifying that
a check failed in \code{A()}.

The call-stack in check is reversed, meaning that
it starts at the check call (0) and goes up the
call stack progressively with \emph{cs_level}.
}
