% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_blanks.R
\name{check_blanks}
\alias{check_blanks}
\title{Checking fill-in-the-blanks problems}
\usage{
check_blanks(ex, pat, ...)
}
\arguments{
\item{ex}{the expression to check, e.g. USER_CODE}

\item{pat}{the anticipated correct form. Use elements such as \code{..one..} to
define a blank.}

\item{...}{tests to apply to expressions in \code{ex}. These are typically made
with \code{passif()}, \code{failif()}, and so on.}
}
\description{
Checking fill-in-the-blanks problems
}
\examples{
submission <- for_checkr(quote({a <- 3; b <- 4; res <- sqrt(a^2 + b^2)}))
submission2 <- for_checkr(quote({a <- 3; b <- 4; res <- sin(a^2 + b^2)}))
# a template with a blank ..fn..
as_posted <- quote(res <- ..fn..(a^2 + b^2))
check_blanks(submission, !!as_posted,
   insist(fn == quote(sqrt), "{{fn}} is not the right function."))
check_blanks(submission2, !!as_posted,
   insist(fn == quote(sqrt), "{{fn}} is not the right function."))
# Multiple blanks
as_posted <- for_checkr(quote({res <- ..fn..(`+`(`^`(a, ..exp1..), `^`(b, ..exp2..)))}))
}
