% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{check_count}
\alias{check_count}
\title{Check Count}
\usage{
check_count(x, coerce = FALSE, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{coerce}{A flag indicating whether to coerce a non-negative numeric (dbl) whole number to a count and drop attributes (including names).}

\item{x_name}{A string of the name of the object x.}

\item{error}{A flag indicating whether to throw an informative error or
immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if an object is a count (non-negative integer or if \code{coerce = TRUE} non-negative numeric whole number).
}
\examples{
check_count(-1L, error = FALSE)
check_count(1L, error = FALSE)
check_count(1, error = FALSE)
check_count(1, coerce = TRUE, error = FALSE)
check_count(1.01, coerce = TRUE, error = FALSE)
}
\seealso{
\code{\link{check_scalar}}
}
