% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{check_is_integerish}
\alias{check_is_integerish}
\alias{check_is_numeric}
\title{Check if a variable is integerish}
\usage{
check_is_integerish(
  x,
  .x_name = rlang::caller_arg(x),
  .call = rlang::caller_env(),
  .class = .error_class
)

check_is_numeric(
  x,
  .x_name = rlang::caller_arg(x),
  .call = rlang::caller_env(),
  .class = .error_class
)
}
\arguments{
\item{x}{the variable to check}

\item{.x_name}{the name of x in the parent environment}

\item{.call}{The caller enviroment, used to get the name of the function that called the check}

\item{.class}{The class of the error, defaults to 'checkr_error'}
}
\value{
TRUE invisibly if the variable is integerish
}
\description{
Check if a variable is integerish
}
\section{Functions}{
\itemize{
\item \code{check_is_numeric()}: check that a variable is numeric

}}
