% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{check_length}
\alias{check_length}
\title{Check Length}
\usage{
check_length(x, length = TRUE, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{length}{A flag indicating whether x should have elements (versus no elements) or a missing value indicating no requirements or a count or count range of the number of elements or a count vector of the permitted number of elements.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether the number of elements in an object is an exact number, 
within a range or 0 vs positive.
}
\examples{
check_length(2)
check_length(character(0), length = 0)
check_length(NULL, error = FALSE)
check_length(list(), error = FALSE)
}
\seealso{
\code{\link{check_vector}}, \code{\link{check_list}} and \code{\link{check_data}}
}
