% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{check_list}
\alias{check_list}
\title{Check List}
\usage{
check_list(x, values, min_length = 0L, max_length = .Machine$integer.max,
  named = NA, x_name = substitute(x),
  error = getOption("vectorcheckr.error", TRUE))
}
\arguments{
\item{x}{The object to check.}

\item{values}{An optional vector specifying the values.}

\item{min_length}{A count of the minimum length.}

\item{max_length}{A count of the maximum length.}

\item{named}{A flag (or NA) indicating whether the list must be named or unnamed (or doesn't matter).}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check List
}
\examples{
x <- list(x = 1, y = 2)
check_list(1:2, error = FALSE)
check_list(list(x = 1, y = 2), named = TRUE, error = FALSE)
}
