% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{check_list}
\alias{check_list}
\title{Check List}
\usage{
check_list(x, values = NULL, exclusive = FALSE, order = FALSE,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{values}{An optional vector or named list specifying the values.}

\item{exclusive}{A flag indicating whether other elements are not permitted.}

\item{order}{A flag indicating whether the elements have to occur in the same order as values.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks whether an object is a list and optionally the names and values
of its elements.
}
\examples{
check_list(list())
check_list(list(x1 = 2, x2 = 1:2), values = list(x1 = 1, x2 = 1L))
}
\seealso{
\code{\link{check_length}} and \code{\link{check_unique}}
}
