% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{check_logit}
\alias{check_logit}
\title{check_logit}
\usage{
check_logit(model)
}
\arguments{
\item{model}{a logit model object to test}
}
\value{
A list containing: the difference between the null deviance and model deviance; the difference in degrees of freedom; the chi-squared probability of the deviance statistic; Cox and Snell pseudo-R^2; Nagelkerke pseudo-R^2; and Hosmer pseudo-R^2.
}
\description{
Check logistic regression statistics and assumptions
}

