% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{check_nrow}
\alias{check_nrow}
\title{Check nrow}
\usage{
check_nrow(x, nrow = c(1L, 2147483647L), x_name = substitute(x),
  error = getOption("checkr.error", TRUE))
}
\arguments{
\item{x}{The data to check.}

\item{nrow}{A count or count range of the number of rows.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check nrow
}
\examples{
check_nrow(data.frame(x = 1), error = FALSE)
check_nrow(data.frame(x = integer(0)), error = FALSE)
check_nrow(data.frame(x = 1:2), nrow = 1, error = FALSE)
}
