% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{check_vector}
\alias{check_vector}
\title{Check Vector}
\usage{
check_vector(x, values, min_length = 0L, max_length = .Machine$integer.max,
  unique = FALSE, sorted = FALSE, named = NA, x_name = substitute(x),
  error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{values}{An optional vector specifying the values.}

\item{min_length}{A count of the minimum length.}

\item{max_length}{A count of the maximum length.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{named}{A flag (or NA) indicating whether the vector must be named or unnamed (or doesn't matter).}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Vector
}
\examples{
check_vector(2:1, min_length = 3, sorted = TRUE, named = TRUE, error = FALSE)
check_vector(c("one", "two", "four"), values = c("one", "two", "two"), error = FALSE)
}
