% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkr-package.R
\docType{package}
\name{checkr-package}
\alias{checkr}
\alias{checkr-package}
\title{checkr: Check the Properties of Common R Objects}
\description{
Expressive, assertive, pipe-friendly functions 
  to check the properties of common R objects.
  In the case of failure the functions issue informative error messages.
}
\seealso{
\code{\link{check_vector}}, \code{\link{check_list}} and \code{\link{check_data}}
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
