% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkr.R
\name{Checkr}
\alias{Checkr}
\title{Class for combining check() function with informative message}
\usage{
Checkr(isvalid, message)
}
\arguments{
\item{isvalid}{isvalid(x) = TRUE if x qualifies, FALSE otherwise}

\item{message}{message(x) returns message if x is not valid}
}
\description{
Class for combining check() function with informative message
}
