% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{\%same_as\%}
\alias{\%same_as\%}
\alias{\%not_same_as\%}
\alias{\%not_same_as\%}
\title{Comparison operators for bindings}
\usage{
e1 \%same_as\% e2

e1 \%not_same_as\% e2
}
\arguments{
\item{e1}{an expression,}

\item{e2}{another expression}
}
\description{
Comparison operators for bindings
}
