% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure}
\alias{ensure}
\title{Ensure checks that certain preconditions and postconditions of a function are true.}
\usage{
ensure(fn, preconditions = list(), postconditions = list())
}
\arguments{
\item{fn}{function. A function to run with validated pre- and postconditions.}

\item{preconditions}{list. A list of preconditions to check.}

\item{postconditions}{list. A list of postconditions to check.}
}
\value{
The original function, but also of class validated_function, with added validations.
}
\description{
Ensure checks that certain preconditions and postconditions of a function are true.
}
\examples{
  add <- ensure(pre = list(x \%is\% numeric, y \%is\% numeric),
    post = list(result \%is\% numeric),
    function(x, y) { x + y })
}

