% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_check.R
\name{error_on_false}
\alias{error_on_false}
\title{check helper}
\usage{
error_on_false(
  check,
  x,
  var_name,
  input_name,
  allow_null,
  allow_na,
  expected,
  call,
  class
)
}
\arguments{
\item{check}{boolean statement to check}

\item{x}{the variable to check}

\item{var_name}{the variable name the user gave the input}

\item{input_name}{the name of the input we are checking}

\item{expected}{a string describing what the input should be}

\item{call}{the call environment}

\item{class}{the class of the error}
}
\value{
invisible true/falses
}
\description{
check helper
}
