% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkr.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a Checkr object}
\usage{
evaluate(chckr, x)
}
\arguments{
\item{chckr}{Checkr object}

\item{x}{object to check}
}
\value{
NULL or character
}
\description{
Evaluate executes the provided isvalid() function of a given Checkr object.
If it returns TRUE, NULL is returned, otherwiese a string generated from
message().
}
