% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_checkr.R
\name{for_checkr}
\alias{for_checkr}
\title{Translate a sequence of commands into a sequence of quosures}
\usage{
for_checkr(exprs)
}
\arguments{
\item{exprs}{an expression or set of expressions in a curly brace}
}
\value{
a quosure for each expression in exs, with an associated
environment that reflects the state when that expression was evaluated.
}
\description{
This function is the first step in any checkr sequence. It scans the
quoted commands for parse errors, evaluates the lines one at a time, and
packages up the lines into quosures with an environment in which the line
should be executed. (That way the lines can be evaluated in a stand-alone
manner.)
}
\examples{
code <- for_checkr(quote({x <- 2; y <- x^2; z <- x + y}))
class(code)
passed(code)
line_where(code, Z == "z")
line_where(code, V == 4, Z == "y")
}
