% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/empty.R
\name{is.empty}
\alias{is.empty}
\title{Tests whether an object is empty.}
\usage{
is.empty(obj)
}
\arguments{
\item{obj.}{The object to test.}
}
\value{
a boolean whether or not the object is empty.
}
\description{
Empty items are NULL, NA, or nothing (length 0).
}
\examples{
is.empty(NULL)
  is.empty(NA)
  is.empty(list(NULL, NA))
  is.empty(list())
  is.empty(c())
  is.empty(data.frame())
  is.empty("")
  is.empty(data.frame())
}

