% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_tidy_table}
\alias{is_tidy_table}
\title{Checks if input is a \code{\link[tibble]{data_frame}}}
\usage{
is_tidy_table(input)
}
\arguments{
\item{input}{The input to check}
}
\description{
Checks if input is a \code{\link[tibble]{data_frame}}
}
