% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_where.R
\name{line_where}
\alias{line_where}
\alias{lines_after}
\alias{lines_after}
\title{Identify an individual line in a sequence of commands}
\usage{
line_where(ex, ..., message = "")

lines_after(ex, ..., message = "")
}
\arguments{
\item{ex}{expressions as made by for_checkr()}

\item{...}{tests specifying the kind of line we want}

\item{message}{A character string to be included as the message in the result. This
can have moustaches written in terms of F, Z, V, or EX}
}
\value{
A checkr test result. By default, if the line is found, the result
is an "OK", setting the stage for further testing. If no matching line is found, the result is a fail.
}
\description{
The tests are written in terms of pronouns
\itemize{
\item F the function at the highest level (but with assignment removed)
\item Z the name being bound to the line's value by assignment. ("" if no assignment.)
\item V the value produced by the line.
\item EX the expression itself (but with assignment removed)
}

Grab the lines after a specified line (which is included)
}
\examples{
ex <- for_checkr(quote({x <- 2; y <- x^3; z <- y + x}))
line_where(ex, F == "^")

}
