% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.checkr_result.R
\name{print.checkr_result}
\alias{print.checkr_result}
\title{Print a checkr_result}
\usage{
\method{print}{checkr_result}(x, ...)
}
\arguments{
\item{x}{a checkr_result object to be printed}

\item{...}{additional args (not used here)}
}
\description{
Print a checkr_result
}
