% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{random_objs}
\alias{random_objs}
\title{Generate a vector or list of random objects from a particular set of possible choices.}
\usage{
random_objs(objects, amount)
}
\description{
Generate a vector or list of random objects from a particular set of possible choices.
}

