% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{random_objs}
\alias{random_objs}
\title{Generate a vector or list of random objects from a particular set of possible choices.}
\usage{
random_objs(objects, amount)
}
\arguments{
\item{objects}{list. The list of objects to generate from.}

\item{amount}{numeric. The amount of objects to generate.}
}
\description{
Generate a vector or list of random objects from a particular set of possible choices.
}

