% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{random_simple_strings}
\alias{random_simple_strings}
\title{Generate a random simple string (i.e., a length-1 non-empty vector of characters).}
\usage{
random_simple_strings(amount, chars = TRUE, utf8 = FALSE)
}
\arguments{
\item{amount}{numeric. The amount of simple strings to generate.}

\item{chars}{logical. Whether or not to include characters.}

\item{utf8}{logical. Whether or not to include utf8 characters.}
}
\description{
Generate a random simple string (i.e., a length-1 non-empty vector of characters).
}

