% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate checks that certain facts are true.}
\usage{
validate(...)
}
\arguments{
\item{...}{list. A list of conditions to check.}
}
\value{
Either TRUE or stops with a list of errors.
}
\description{
Validate checks that certain facts are true.
}
\examples{
  validate(1 == 1, "a" \%is\% character, length(c(1, 2, 3)) == 3)
}
